move MapTarget to separate package
